backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Redis for caching and pub/sub
redis==5.0.1
aioredis==2.0.1

# Web3 / Blockchain
web3==6.14.0
eth-account==0.10.0
eth-typing==3.5.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# WebSocket and real-time
websockets==12.0

# HTTP Client for external APIs
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
orjson==3.9.12
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

